+++ packages/tools/utils/src/object-[object-object].spec.js
@@ -2,30 +2,65 @@ import {strict as assert} from 'assert';
 
 import * as _ from 'lamb';
 
+import {roundTo} from './number-[number-number]';
 import {joinWith} from './string-[array-string]';
 import {
 	applyFnMap,
+	makeMergeAppliedFnMap,
 	makeMergeKeyValue,
 	mergeObj,
 	transformPaths,
 	transformValues,
 } from './object-[object-object]';
 
+const roundTo2 = roundTo(2);
+
 describe('Object -> (Object -> Object)', function() {
 	describe('applyFnMap', function() {
 		it('should return a function expecting an object to be used as the argument of the provided functions', function() {
+			const object = {fname: 'John', lname: 'Woo', lng: 1, lat: 2};
 			const format = applyFnMap({
+				coords: _.collect([_.getKey('lng'), _.getKey('lat')]),
 				fullname: _.pipe([
 					_.collect([_.getKey('fname'), _.getKey('lname')]),
 					joinWith(' ')
 				]),
-				coords: _.collect([_.getKey('lng'), _.getKey('lat')])
 			});
 
 			assert.deepStrictEqual(
-				format({fname: 'John', lname: 'Woo', lng: 1, lat: 2}),
-				{fullname: 'John Woo', coords: [1, 2]}
+				format(object),
+				{coords: [1, 2], fullname: 'John Woo'}
 			);
+		});
+	});
+	describe('makeMergeAppliedFnMap', function() {
+		it('should return a function that applies the provided map to the expected object and merges te result to the object', function() {
+			const enhancer = makeMergeAppliedFnMap({
+				coords: _.collect([_.getKey('lng'), _.getKey('lat')]),
+				fullname: _.pipe([
+					_.collect([_.getKey('fname'), _.getKey('lname')]),
+					joinWith(' ')
+				]),
+				lat: obj => roundTo2(obj.lat),
+				lng: obj => roundTo2(obj.lng),
+			});
+			const object = {
+				fname: 'John',
+				lat: 2.345434,
+				lname: 'Woo',
+				lng: 10.3425,
+			};
+			const expected = {
+				coords: [10.3425, 2.345434],
+				fname: 'John',
+				fullname: 'John Woo',
+				lat: 2.35,
+				lname: 'Woo',
+				lng: 10.34,
+			};
+			const actual = enhancer(object);
+
+			assert.deepStrictEqual(actual, expected);
 		});
 	});